# microservices/service-a/pkg/adapter/http/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Desafio Sistema de Temperatura por Cep Go Expert API Docs",
        "contact": {
            "name": "Vinícius Boscardin",
            "email": "boscardinvinicius@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/cep/{cep}": {
            "get": {
                "description": "Get temperature in celcius, kelvin and fahrenheit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cep"
                ],
                "summary": "Get temperature",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cep",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WeatherOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorhandle.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorhandle.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.WeatherOutput": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "temp_C": {
                    "type": "number"
                },
                "temp_F": {
                    "type": "number"
                },
                "temp_K": {
                    "type": "number"
                }
            }
        },
        "errorhandle.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}